# src/main/resources/static/openapi.yml
openapi: 3.0.3
info:
  title: RLibrary Backend API
  version: 1.0.0
  description: REST API for managing books, authors, and loans in the RLibrary system.
security:
  - auth0_jwt: []
servers:
  - url: /api
  - url: http://localhost:8080/api
paths:
...
  /libraries:
    get:
      summary: Get all libraries
      operationId: librariesGet
      tags:
        - Library
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            default: 20
        - in: query
          name: sort
          schema:
            type: string
            default: name
          description: Field to sort by (id, name, city)
        - in: query
          name: direction
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: Sort direction
        - in: query
          name: name
          schema:
            type: string
          description: Filter by name (contains, case-insensitive)
        - in: query
          name: city
          schema:
            type: string
          description: Filter by city (contains, case-insensitive)
      responses:
        '200':
          description: List of libraries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedLibraryResponse'
    post:
      summary: Create a new library
      operationId: librariesPost
      tags:
        - Library
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LibraryDto'
      responses:
        '201':
          description: Library created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibraryDto'
        '409':
          description: A library with this name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /libraries/{id}:
    get:
      summary: Get library by ID
      operationId: librariesIdGet
      tags:
        - Library
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Library found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibraryDto'
        '404':
          description: Library not found
    put:
      summary: Update library by ID
      operationId: librariesIdPut
      tags:
        - Library
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LibraryDto'
      responses:
        '200':
          description: Library updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibraryDto'
        '404':
          description: Library, LibraryRule or LibraryStats not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete library by ID
      operationId: librariesIdDelete
      tags:
        - Library
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Library deleted
        '404':
          description: Library not found
  /libraries/{id}/books:
    get:
      summary: Get books for a library
      operationId: librariesIdBooksGet
      tags:
        - Library
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            default: 20
      responses:
        '200':
          description: List of books in the library
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedBookResponse'
        '404':
          description: Library not found
  /loans:
    get:
      summary: Get all loans
      operationId: loansGet
      tags:
        - Loan
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            default: 20
        - in: query
          name: sort
          schema:
            type: string
            default: loanDate
          description: Field to sort by (id, loanDate, returnDate, status)
        - in: query
          name: direction
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: Sort direction
        - in: query
          name: status
          schema:
            type: string
          description: Filter by loan status
        - in: query
          name: personId
          schema:
            type: integer
            format: int64
          description: Filter by person id
        - in: query
          name: loanDateFrom
          schema:
            type: string
            format: date
          description: Filter loans from this date (inclusive)
        - in: query
          name: loanDateTo
          schema:
            type: string
            format: date
          description: Filter loans until this date (inclusive)
      responses:
        '200':
          description: List of loans
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedLoanResponse'
    post:
      summary: Create a new loan
      operationId: loansPost
      tags:
        - Loan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanDto'
      responses:
        '201':
          description: Loan created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanDto'
  /loans/{id}:
    get:
      summary: Get loan by ID
      operationId: loansIdGet
      tags:
        - Loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Loan found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanDto'
        '404':
          description: Loan not found
    put:
      summary: Update loan by ID
      operationId: loansIdPut
      tags:
        - Loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanDto'
      responses:
        '200':
          description: Loan updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanDto'
        '404':
          description: Loan, Person, LoanRule or Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete loan by ID
      operationId: loansIdDelete
      tags:
        - Loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Loan deleted
        '404':
          description: Loan not found
  /loans/book/{bookId}:
    get:
      summary: Get all loans for a specific book
      operationId: loansBookBookIdGet
      tags:
        - Loan
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            default: 20
        - in: query
          name: sort
          schema:
            type: string
            default: loanDate
          description: Field to sort by (id, loanDate, returnDate, status)
        - in: query
          name: direction
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort direction
      responses:
        '200':
          description: List of loans for the book
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedLoanResponse'
        '404':
          description: Book not found
...
components:
  securitySchemes:
    auth0_jwt:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://dev-m362yc372ygvz1ny.us.auth0.com/authorize?audience=https://rlibrary.com
          tokenUrl: https://dev-m362yc372ygvz1ny.us.auth0.com/oauth/token
          scopes:
            openid: OpenId Connect
            profile: User Profile
            email: User Email
  schemas:
...
    BookDto:
      type: object
      required:
        - title
        - isbn
        - authorId
        - libraryId
        - publishedYear
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        authorName:
          type: string
          readOnly: true
        isbn:
          type: string
          minLength: 10
          maxLength: 17
        authorId:
          type: integer
          format: int64
        libraryId:
          type: integer
          format: int64
        libraryName:
          type: string
          readOnly: true
        publishedYear:
          type: integer
          minimum: 1400
          maximum: 2100
        amountOfPages:
          type: integer
          minimum: 1
          maximum: 10000
        releaseDate:
          type: string
          format: date
        theme:
          $ref: '#/components/schemas/ThemeType'
        bookState:
          $ref: '#/components/schemas/BookState'
        ageCategory:
          type: string
          maxLength: 50
        purchasePrice:
          type: number
          format: double
          minimum: 0.0
        duplicates:
          type: integer
          minimum: 0
        coverImage:
          type: string
          format: byte
          description: Base64 encoded cover image
    BookState:
      type: string
      enum:
        - AVAILABLE
        - TEMPORARILYUNAVAILABLE
        - PERMANENTLYUNAVAILABLE
        - LOST
        - BORROWED
    LoanDto:
      type: object
      required:
        - loanDate
        - personId
        - bookIds
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        loanDate:
          type: string
          format: date-time
        returnDate:
          type: string
          format: date-time
        status:
          type: string
        personId:
          type: integer
          format: int64
        personName:
          type: string
          readOnly: true
          description: Full name of the person who made the loan (firstName + lastName)
        loanRuleId:
          type: integer
          format: int64
        bookIds:
          type: array
          minItems: 1
          items:
            type: integer
            format: int64
...
    ErrorResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: integer
        message:
          type: string
    LibraryDto:
      type: object
      required:
        - name
        - country
        - city
        - street
        - streetNumber
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        name:
          type: string
          minLength: 1
          maxLength: 200
        country:
          type: string
          minLength: 1
          maxLength: 100
        city:
          type: string
          minLength: 1
          maxLength: 100
        street:
          type: string
          minLength: 1
          maxLength: 100
        streetNumber:
          type: string
          minLength: 1
          maxLength: 20
        description:
          type: string
          maxLength: 1000
        address:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
        website:
          type: string
        openingHours:
          type: string
        libraryStatsId:
          type: integer
          format: int64
        libraryRuleId:
          type: integer
          format: int64
        bookIds:
          type: array
          items:
            type: integer
            format: int64
...
    ThemeType:
      type: string
      enum:
        - FINANCE
        - LAWS
        - COMMUNICATION
        - BUSINESS
        - COMPUTERSCIENCE
        - IT
...
    PagedBookResponse:
      type: object
      required:
        - page
        - size
        - totalElements
        - totalPages
        - items
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          minimum: 0
        items:
          type: array
          items:
            $ref: '#/components/schemas/BookDto'
    PagedLibraryResponse:
      type: object
      required:
        - page
        - size
        - totalElements
        - totalPages
        - items
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          minimum: 0
        items:
          type: array
          items:
            $ref: '#/components/schemas/LibraryDto'
    PagedLoanResponse:
      type: object
      required:
        - page
        - size
        - totalElements
        - totalPages
        - items
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          minimum: 0
        items:
          type: array
          items:
            $ref: '#/components/schemas/LoanDto'
...